% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\name{goodbye}
\alias{goodbye}
\title{My goodbye function}
\usage{
goodbye(x)
}
\arguments{
\item{x}{The name of the person to say goodbye to}
}
\value{
The output from \code{link{print}}
}
\description{
My goodbye function
}
\examples{
goodbye("Covid")
\dontrun{
goodbye("Fun")
}
}
